+++ fs/sysfs/mounts.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+set -e
+
+dir=$(mktemp --directory sysfs-XXXXXX)
+
+cleanup() {
+	sudo umount ${dir} || true
+	sudo rmdir ${dir}
+}
+trap cleanup EXIT
+
+sudo mount -t sysfs sysfs ${dir}
+ls ${dir}/